updated code by olaer and cruzado (ft. bolanos)
defect 1 - no instruction about month was specified; a user might enter the name of the month instead of entering the number of the month (i.e. January is 1, June is 6, etc.) [ADDED]

defect 2 - in lines 44 and 46, the numbers are directly added instead of being separately shown. [CORRECTED]

defect 3 -  there is no condition for leap year [ADDED]

defect 4 - despite there being no line of code to indicate for janurary [(month == 1) or case: 1], the code somehow still works and runs as if a code for january exists. [ADDED]

defect 5 - lines 41 to 65 are using if-else statements [FIXED]

defect 6 - line spacing is not necessary [FIXED]

defect 7 - unnecessary parsing in line 20 - 21 [FIXED]