Update Main.java
Etang & Alghanim

CHANGES: I just added another method which utilize to determines whether a given year is a leap year or not. because in a leap year, Feb has 29 days instead of 28 which is usually the days in Feb. The method "isLeapYear" takes an integer parameter year representing the year to be checked and it returns a boolean value (true if it's a leap year, false otherwise).

The added code:
  "public static boolean isLeapYear(int year) {
        return (year % 4 == 0 && year % 100 !=0) || (year % 400 == 0);
    }
}"