Update Main.java
Etang & Alghanim

Changes: I made a method called "isLeapYear" which determines the given year is a leap year or not. And method also applies a parameter, an integer "year" representing the year to be checked and returns a boolean value (true if it's a leap year, false otherwise). Plus the days that was added on the if else statement that was on the code was change to it's exact value and not added every time.